% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_discrete_cosine_transoformation.R
\name{DCT2D}
\alias{DCT2D}
\title{Two Dimensional Discrete Cosine Transformation}
\usage{
DCT2D(imormat, returnmat = FALSE)
}
\arguments{
\item{imormat}{a grayscale image of class cimg or a numeric matrix}

\item{returnmat}{if returnmat is TRUE, DCT2D returns numeric matrix. if FALSE, DCT2D returns a grayscale image of class cimg.}
}
\value{
a grayscale image of class cimg or a numeric matrix
}
\description{
compute two dimensional discrete cosine transformation
}
\examples{
g <- grayscale(boats)
DCT2D(g)
}
\references{
Makhoul, J. (1980). A fast cosine transform in one and two dimensions. IEEE Transactions on Acoustics, Speech, and Signal Processing. 28 (1): 27-34.
}
\author{
Shota Ochi
}
