% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dogs}
\alias{dogs}
\title{Photograph of a dog from GAHAG}
\format{an image of class cimg}
\source{
\url{http://gahag.net/img/201603/03s/gahag-0062116383-1.jpg}
}
\usage{
dogs
}
\description{
This photograph was downloaded from http://gahag.net/img/201603/03s/gahag-0062116383-1.jpg.
Its size was reduced by half to speed up loading and save space.
}
\keyword{datasets}
