% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_color_image.R
\name{GetHue}
\alias{GetHue}
\title{store hue of color image}
\usage{
GetHue(imcol)
}
\arguments{
\item{imcol}{a color image of class cimg}
}
\value{
a color image of class cimg
}
\description{
store hue of color image
}
\examples{
GetHue(boats)
}
\author{
Shota Ochi
}
