% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_discrete_cosine_transoformation.R
\name{IDCT2D}
\alias{IDCT2D}
\title{Two Dimensional Inverse Discrete Cosine Transformation}
\usage{
IDCT2D(imormat, returnmat = FALSE)
}
\arguments{
\item{imormat}{a grayscale image of class cimg or a numeric matrix}

\item{returnmat}{if returnmat is TRUE, IDCT2D returns numeric matrix. if FALSE, IDCT2D returns a grayscale image of class cimg.}
}
\value{
a grayscale image of class cimg or a numeric matrix
}
\description{
compute two dimensional inverse discrete cosine transformation
}
\examples{
g <- grayscale(boats)
layout(matrix(1:4, 2, 2))
plot(boats, main = "Original")
plot(boats, main = "Monochrome")
gg <- DCT2D(g) \%>\% plot(main = "Transformed")
IDCT2D(gg) \%>\% plot(main = "Retransformed")
}
\references{
Makhoul, J. (1980). A fast cosine transform in one and two dimensions. IEEE Transactions on Acoustics, Speech, and Signal Processing. 28 (1): 27-34.
}
\author{
Shota Ochi
}
