% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{package}
\name{imagerExtra}
\alias{imagerExtra}
\title{imagerExtra: Extra Image Processing Library Based on Imager}
\description{
imagerExtra is built on imager. imager by Simon Simon Barthelme provides an interface with CImg that is a C++ library for image processing. imager makes functions of CImg accessible from R and adds many utilities for accessing and working with image data from R.
imagerExtra provides advanced functions for image processing based on imager.
}
