% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_color_image.R
\name{RestoreHue}
\alias{RestoreHue}
\title{restore hue of color image}
\usage{
RestoreHue(im, hueim)
}
\arguments{
\item{im}{a grayscale image of class cimg}

\item{hueim}{a color image of class cimg}
}
\value{
a color image of class cimg
}
\description{
restore hue of color image
}
\examples{
g <- Grayscale(boats)
hue <- GetHue(boats)
layout(matrix(1:2, 1, 2))
plot(g, main = "Original")
RestoreHue(g, hue) \%>\% plot(main="Resotred")
}
\author{
Shota Ochi
}
