% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_thresholding.R
\name{ThresholdML}
\alias{ThresholdML}
\title{Multilevel Thresholding (Maximum Entropy Based Artificial Bee Colony Thresholding)}
\usage{
ThresholdML(im, k, sn = 30, mcn = 100, limit = 100,
  intervalnumber = 1000, returnvalue = FALSE)
}
\arguments{
\item{im}{a grayscale image of class cimg}

\item{k}{the number of thresholds}

\item{sn}{population size}

\item{mcn}{maximum cycle number}

\item{limit}{abandonment criteria}

\item{intervalnumber}{interval number of histogram}

\item{returnvalue}{if returnvalue is TRUE, returns threshold values. if FALSE, returns a grayscale image of class cimg.}
}
\value{
a grayscale image of class cimg or a numeric vector
}
\description{
automatic multilevel thresholding based on Maximum Entropy Based Artificial Bee Colony Thresholding
}
\examples{
g <- grayscale(boats)
ThresholdML(g, 2) \%>\% plot
}
\references{
Ming-HuwiHorng (2011). Multilevel thresholding selection based on the artificial bee colony algorithm for image segmentation. Expert Systems with Applications.
}
\author{
Shota Ochi
}
