% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCR.R
\name{OCR}
\alias{OCR}
\alias{OCR_data}
\title{Optical Character Recognition with tesseract}
\usage{
OCR(imorpx, engine = tesseract::tesseract("eng"), HOCR = FALSE)

OCR_data(imorpx, engine = tesseract::tesseract("eng"))
}
\arguments{
\item{imorpx}{a grayscale image of class cimg or a pixel set}

\item{engine}{a tesseract engine. See the reference manual of tesseract for detail.}

\item{HOCR}{if TRUE return results as HOCR xml instead of plain text}
}
\description{
OCR and OCR_data are wrappers for ocr and ocr_data of tesseract package.
You need to install tesseract package to use these functions.
}
\examples{
hello <- DenoiseDCT(papers, 0.01) \%>\% ThresholdAdaptive(., 0.1, range = c(0,1))
if (requireNamespace("tesseract", quietly = TRUE))
{
  OCR(hello) \%>\% cat
  OCR_data(hello)
}
}
\author{
Shota Ochi
}
